deepcode-mcp/deepcode_lab/papers/1/generate_code/ai_tech_web/requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
requests>=2.28.0
streamlit-extras>=0.3.0